+++ src/main/java/com/powsybl/openloadflow/DefaultOuterLoopConfig.java
@@ -8,6 +8,8 @@ package com.powsybl.openloadflow;
 
 import com.powsybl.loadflow.LoadFlowParameters;
 import com.powsybl.openloadflow.ac.outerloop.*;
+import com.powsybl.openloadflow.lf.outerloop.OuterLoop;
+import com.powsybl.openloadflow.lf.outerloop.OuterLoopConfig;
 import com.powsybl.openloadflow.network.util.ActivePowerDistribution;
 
 import java.util.ArrayList;
