+++ src/main/java/com/powsybl/openloadflow/ac/AcLoadFlowResult.java
@@ -6,7 +6,7 @@
  */
 package com.powsybl.openloadflow.ac;
 
-import com.powsybl.openloadflow.OuterLoopStatus;
+import com.powsybl.openloadflow.lf.outerloop.OuterLoopStatus;
 import com.powsybl.openloadflow.lf.AbstractLoadFlowResult;
 import com.powsybl.openloadflow.ac.nr.NewtonRaphsonStatus;
 import com.powsybl.openloadflow.network.LfNetwork;
