+++ src/main/java/com/powsybl/openloadflow/lf/outerloop/OuterLoop.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.openloadflow;
+package com.powsybl.openloadflow.lf.outerloop;
 
 import com.powsybl.commons.reporter.Reporter;
