+++ src/main/java/com/powsybl/openloadflow/lf/outerloop/OuterLoopContext.java
@@ -4,12 +4,20 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.openloadflow;
+package com.powsybl.openloadflow.lf.outerloop;
+
+import com.powsybl.openloadflow.network.LfNetwork;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public enum OuterLoopStatus {
-    STABLE,
-    UNSTABLE
+public interface OuterLoopContext {
+
+    LfNetwork getNetwork();
+
+    int getIteration();
+
+    Object getData();
+
+    void setData(Object data);
 }
