+++ src/main/java/com/powsybl/openloadflow/lf/outerloop/OuterLoopStatus.java
@@ -4,20 +4,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.openloadflow;
-
-import com.powsybl.openloadflow.network.LfNetwork;
+package com.powsybl.openloadflow.lf.outerloop;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public interface OuterLoopContext {
-
-    LfNetwork getNetwork();
-
-    int getIteration();
-
-    Object getData();
-
-    void setData(Object data);
+public enum OuterLoopStatus {
+    STABLE,
+    UNSTABLE
 }
